LICTCLUB/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-anymail==13.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
ecdsa==0.19.1
gunicorn==21.2.0
idna==3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.6.0
